vlc-video: Fix sign-compare warning

+++ plugins/vlc-video/vlc-video-source.c
@@ -509,7 +509,7 @@ static int vlcs_audio_setup(void **p_data, char *format, unsigned *rate,
 	enum audio_format new_audio_format;
 	struct obs_audio_info aoi;
 	obs_get_audio_info(&aoi);
-	int out_channels = (int)get_audio_channels(aoi.speakers);
+	uint32_t out_channels = get_audio_channels(aoi.speakers);
 
 	new_audio_format = convert_vlc_audio_format(format);
 	if (*channels > out_channels)
